requirements.txt
playwright==1.40.0
pytesseract==0.3.10
Pillow==10.1.0
requests==2.31.0
opencv-python==4.8.1.78
flask==3.0.0
python-dateutil==2.8.2
beautifulsoup4==4.12.2
lxml==4.9.3
